Labs: Remnant files and solved stages #208   - Renamed all redefined GoatHills lesson classes to a name which reflects the lesson to stop confusing the imports.
