+++ sql-injection/src/main/java/org/owasp/webgoat/plugin/sqlinjection/ViewProfileSqlInjection.java
@@ -1,5 +1,5 @@
 
-package org.owasp.webgoat.plugin.rollbased;
+package org.owasp.webgoat.plugin.sqlinjection;
 
 import org.owasp.webgoat.plugin.GoatHillsFinancial.DefaultLessonAction;
 import org.owasp.webgoat.plugin.GoatHillsFinancial.Employee;
@@ -9,9 +9,9 @@ import org.owasp.webgoat.session.UnauthenticatedException;
 import org.owasp.webgoat.session.UnauthorizedException;
 import org.owasp.webgoat.session.WebSession;
 
-import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.sql.Statement;
 
 
 /***************************************************************************************************
@@ -41,10 +41,10 @@ import java.sql.SQLException;
  * 
  * For details, please see http://webgoat.github.io
  */
-public class EditProfile extends DefaultLessonAction
+public class ViewProfileSqlInjection extends DefaultLessonAction
 {
 
-    public EditProfile(GoatHillsFinancial lesson, String lessonName, String actionName)
+    public ViewProfileSqlInjection(GoatHillsFinancial lesson, String lessonName, String actionName)
     {
         super(lesson, lessonName, actionName);
     }
@@ -54,37 +54,59 @@ public class EditProfile extends DefaultLessonAction
     {
         getLesson().setCurrentAction(s, getActionName());
 
+        Employee employee = null;
+
         if (isAuthenticated(s))
         {
-            int userId = getUserId(s);
-            int employeeId = s.getParser().getIntParameter(RoleBasedAccessControl.EMPLOYEE_ID);
-            Employee employee = getEmployeeProfile(s, userId, employeeId);
-            setSessionAttribute(s, getLessonName() + "." + RoleBasedAccessControl.EMPLOYEE_ATTRIBUTE_KEY, employee);
+            String userId = getSessionAttribute(s, getLessonName() + "." + org.owasp.webgoat.plugin.sqlinjection.SQLInjection.USER_ID);
+            String employeeId = null;
+            try
+            {
+                // User selected employee
+                employeeId = s.getParser().getRawParameter(org.owasp.webgoat.plugin.sqlinjection.SQLInjection.EMPLOYEE_ID);
+            } catch (ParameterNotFoundException e)
+            {
+                // May be an internally selected employee
+                employeeId = getRequestAttribute(s, getLessonName() + "." + org.owasp.webgoat.plugin.sqlinjection.SQLInjection.EMPLOYEE_ID);
+            }
+
+            // FIXME: If this fails and returns null, ViewProfile.jsp will blow up as it expects an
+            // Employee.
+            // Most other JSP's can handle null session attributes.
+            employee = getEmployeeProfile(s, userId, employeeId);
+            // If employee==null redirect to the error page.
+            if (employee == null)
+                getLesson().setCurrentAction(s, org.owasp.webgoat.plugin.sqlinjection.SQLInjection.ERROR_ACTION);
+            else
+                setSessionAttribute(s, getLessonName() + "." + org.owasp.webgoat.plugin.sqlinjection.SQLInjection.EMPLOYEE_ATTRIBUTE_KEY, employee);
         }
         else
             throw new UnauthenticatedException();
+
+        updateLessonStatus(s, employee);
     }
 
     public String getNextPage(WebSession s)
     {
-        return RoleBasedAccessControl.EDITPROFILE_ACTION;
+        return org.owasp.webgoat.plugin.sqlinjection.SQLInjection.VIEWPROFILE_ACTION;
     }
 
-    public Employee getEmployeeProfile(WebSession s, int userId, int subjectUserId) throws UnauthorizedException
+    public Employee getEmployeeProfile(WebSession s, String userId, String subjectUserId) throws UnauthorizedException
     {
         Employee profile = null;
 
         // Query the database for the profile data of the given employee
         try
         {
-            String query = "SELECT * FROM employee WHERE userid = ?";
+            String query = "SELECT employee.* "
+                    + "FROM employee,ownership WHERE employee.userid = ownership.employee_id and "
+                    + "ownership.employer_id = " + userId + " and ownership.employee_id = " + subjectUserId;
 
             try
             {
-                PreparedStatement answer_statement = WebSession.getConnection(s)
-                        .prepareStatement(query, ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_READ_ONLY);
-                answer_statement.setInt(1, subjectUserId);
-                ResultSet answer_results = answer_statement.executeQuery();
+                Statement answer_statement = WebSession.getConnection(s)
+                        .createStatement(ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_READ_ONLY);
+                ResultSet answer_results = answer_statement.executeQuery(query);
                 if (answer_results.next())
                 {
                     // Note: Do NOT get the password field.
@@ -96,6 +118,7 @@ public class EditProfile extends DefaultLessonAction
                                     .getInt("salary"), answer_results.getString("ccn"), answer_results
                                     .getInt("ccn_limit"), answer_results.getString("disciplined_date"), answer_results
                                     .getString("disciplined_notes"), answer_results.getString("personal_description"));
+                    // System.out.println("Profile: " + profile);
                     /*
                      * System.out.println("Retrieved employee from db: " + profile.getFirstName() +
                      * " " + profile.getLastName() + " (" + profile.getId() + ")");
@@ -114,7 +137,8 @@ public class EditProfile extends DefaultLessonAction
         return profile;
     }
 
-    public Employee getEmployeeProfile_BACKUP(WebSession s, int userId, int subjectUserId) throws UnauthorizedException
+    public Employee getEmployeeProfile_BACKUP(WebSession s, String userId, String subjectUserId)
+            throws UnauthorizedException
     {
         // Query the database to determine if this employee has access to this function
         // Query the database for the profile data of the given employee if "owned" by the given
@@ -125,14 +149,13 @@ public class EditProfile extends DefaultLessonAction
         // Query the database for the profile data of the given employee
         try
         {
-            String query = "SELECT * FROM employee WHERE userid = ?";
+            String query = "SELECT * FROM employee WHERE userid = " + subjectUserId;
 
             try
             {
-                PreparedStatement answer_statement = WebSession.getConnection(s)
-                        .prepareStatement(query, ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_READ_ONLY);
-                answer_statement.setInt(1, subjectUserId);
-                ResultSet answer_results = answer_statement.executeQuery();
+                Statement answer_statement = WebSession.getConnection(s)
+                        .createStatement(ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_READ_ONLY);
+                ResultSet answer_results = answer_statement.executeQuery(query);
                 if (answer_results.next())
                 {
                     // Note: Do NOT get the password field.
@@ -160,6 +183,48 @@ public class EditProfile extends DefaultLessonAction
         }
 
         return profile;
+    }
+
+    private void updateLessonStatus(WebSession s, Employee employee)
+    {
+        try
+        {
+            String userId = getSessionAttribute(s, getLessonName() + "." + org.owasp.webgoat.plugin.sqlinjection.SQLInjection.USER_ID);
+            String employeeId = s.getParser().getRawParameter(org.owasp.webgoat.plugin.sqlinjection.SQLInjection.EMPLOYEE_ID);
+            String stage = getStage(s);
+            if (org.owasp.webgoat.plugin.sqlinjection.SQLInjection.STAGE3.equals(stage))
+            {
+                // If the employee we are viewing is the prize and we are not authorized to have it,
+                // the stage is completed
+                if (employee != null && employee.getId() == org.owasp.webgoat.plugin.sqlinjection.SQLInjection.PRIZE_EMPLOYEE_ID
+                        && !isAuthorizedForEmployee(s, Integer.parseInt(userId), employee.getId()))
+                {
+                    setStageComplete(s, org.owasp.webgoat.plugin.sqlinjection.SQLInjection.STAGE3);
+                }
+            }
+            else if (org.owasp.webgoat.plugin.sqlinjection.SQLInjection.STAGE4.equals(stage))
+            {
+                // If we were denied the employee to view, and we would have been able to view it
+                // in the broken state, the stage is completed.
+                // This assumes the student hasn't modified getEmployeeProfile_BACKUP().
+                if (employee == null)
+                {
+                    Employee targetEmployee = null;
+                    try
+                    {
+                        targetEmployee = getEmployeeProfile_BACKUP(s, userId, employeeId);
+                    } catch (UnauthorizedException e)
+                    {
+                    }
+                    if (targetEmployee != null && targetEmployee.getId() == org.owasp.webgoat.plugin.sqlinjection.SQLInjection.PRIZE_EMPLOYEE_ID)
+                    {
+                        setStageComplete(s, org.owasp.webgoat.plugin.sqlinjection.SQLInjection.STAGE4);
+                    }
+                }
+            }
+        } catch (ParameterNotFoundException pnfe)
+        {
+        }
     }
 
 }
